fix : fetching user details on refresh
