+++ packages/app/src/components/User/helper.js
@@ -4,15 +4,6 @@ export const getUserDetails = async () => {
   try {
     const response = await GlitchApi.get("/user");
     return response.data;
-    // const res = await fetch(`${API}/user`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //     Authorization: `Bearer`,
-    //   },
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
@@ -23,14 +14,6 @@ export const updateLikedVideos = async (userDetails) => {
     const response = await GlitchApi.post("/user", {
       ...userDetails,
     });
-    // const res = await fetch(`${API}/user`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //   },
-    //   body: userDetails,
-    // });
     return response.data;
   } catch (error) {
     console.log(error);
@@ -43,55 +26,18 @@ export const signInUser = async (user) => {
       ...user,
     });
     return response.data;
-    // const res = await fetch(`${API}/signin`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //   },
-    //   body: JSON.stringify(user),
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
 };
 
-// export const getLikedVideos = async () => {
-//   try {
-//     const response = await GlitchApi.get("/likedVideos");
-//     return response.data;
-//     // const res = await fetch(`${API}/likedvideos`);
-//     // return res.json();
-//   } catch (error) {
-//     console.log(error);
-//   }
-// };
-
-// export const getHistory = async () => {
-//   try {
-//     const response = await GlitchApi.get("/history");
-//     return response.data;
-//     // const res = await fetch(`${API}/history`);
-//     // return res.json();
-//   } catch (error) {
-//     console.log(error);
-//   }
-// };
-
 export const setLikeVideo = async (videoId) => {
   try {
-    console.log(videoId)
-    console.log(GlitchApi.defaults.headers)
-    // response is not getting back
+    console.log(videoId);
+    console.log(GlitchApi.defaults.headers);
     const response = await GlitchApi.post(`/user/likedvideo/${videoId}`);
-    console.log(response.data,"hehe")
+    console.log(response.data, "hehe");
     return response.data;
-
-    // const res = await fetch(`${API}/user/likedvideo/${videoId}`, {
-    //   method: "post",
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
@@ -103,15 +49,6 @@ export const setPlaylist = async (playlist) => {
       ...playlist,
     });
     return response.data;
-    // const res = await fetch(`${API}/user/playlist`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //   },
-    //   body: JSON.stringify(playlist),
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
@@ -126,15 +63,6 @@ export const removeUserPlaylist = async (playlistId, playlist) => {
       }
     );
     return response.data;
-    // const res = await fetch(`${API}/user/removeplaylist/${playlistId}`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //   },
-    //   body: JSON.stringify(playlist),
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
@@ -144,11 +72,6 @@ export const setHistory = async (videoId) => {
   try {
     const response = await GlitchApi.post(`/user/history/${videoId}`);
     return response.data;
-    // console.log("history read");
-    // const res = await fetch(`${API}/user/history/${videoId}`, {
-    //   method: "post",
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
@@ -160,20 +83,7 @@ export const setSuscription = async (suscription) => {
       ...suscription,
     });
     return response.data;
-    // const res = await fetch(`${API}/user/suscription`, {
-    //   method: "post",
-    //   headers: {
-    //     Accept: "application/json",
-    //     "Content-Type": "application/json",
-    //   },
-    //   body: JSON.stringify(suscription),
-    // });
-    // return res.json();
   } catch (error) {
     console.log(error);
   }
 };
-
-// export const isSignedIn = () => {
-//   return JSON.parse(localStorage.getItem("user"));
-// };
